Moving types to file
